requirements.txt
pydantic
pydantic-settings
python-dotenv
ruff
matplotlib
pandas
jira
openpyxl
atlassian-python-api